Modified   Project/Client/Assets/Libraries/BinarySerializerManaged.dll Modified   Project/Client/Assets/Libraries/Networking.Client.dll Modified   Project/Client/Assets/Libraries/Networking.Common.dll Modified   Project/Client/Assets/Libraries/Simulation.Data.dll Modified   Project/Networking/Admin/Network.cs Modified   Project/Networking/Binaries/BinarySerializerManaged.dll Modified   Project/Networking/Client/Connection.cs Modified   Project/Networking/Client/Network.cs Modified   Project/Networking/Server/Admin.cs Modified   Project/Networking/Server/Lobby.cs Modified   Project/Networking/Server/OneByBotRoom.cs Modified   Project/Networking/Server/OneByOneRoom.cs Modified   Project/Networking/Server/Room.cs Modified   Project/Networking/Server/ServerSocket.cs Modified   Project/Networking/Test/Program.cs Modified   Project/Simulation/Binaries/BinarySerializerManaged.dll Modified   Project/Simulation/Data/Serialization/Deserializer.cs Modified   Project/Simulation/Data/Serialization/SerializerVisitor.cs Modified   Project/Simulation/Data/Serialization/SessionDeserializer.cs

+++ Project/Simulation/Data/Serialization/SerializerVisitor.cs
@@ -27,7 +27,7 @@ namespace Simulation.Data.Serialization
 
 		public void BeginVisitArray(ICollection Collection)
 		{
-			buffer.BeginWriteArray(Collection == null ? 0 : Collection.Count);
+			buffer.BeginWriteArray(Collection == null ? 0 : (uint)Collection.Count);
 		}
 
 		public void EndVisitArray()
